requirements.txt
selenium
chromedriver-binary
requests-html
asyncio
nest_asyncio
